split single action controllers
