+++ packages/Marketing/src/Controller/MarketingSponsoringController.php
@@ -8,12 +8,12 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-final class MarketingController extends AbstractController
+final class MarketingSponsoringController extends AbstractController
 {
     /**
      * @Route(path="/zviditelnete-vasi-firmu", name="sponsoring")
      */
-    public function default(): Response
+    public function run(): Response
     {
         return $this->render('marketing/sponsoring.twig');
     }
