+++ packages/Registration/src/Controller/TrainingRegistrationThankYouController.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pehapkari\Registration\Controller;
+
+use Pehapkari\Training\Entity\TrainingTerm;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+final class TrainingRegistrationThankYouController extends AbstractController
+{
+    /**
+     * @Route(path="/vitej-na-skoleni/{slug}/", name="registration_thank_you")
+     */
+    public function run(TrainingTerm $trainingTerm): Response
+    {
+        return $this->render('registration/thank_you_for_registration.twig', [
+            'trainingTerm' => $trainingTerm,
+        ]);
+    }
+}
